Refactoring and minor test change.
git-svn-id: https://studio.atlassian.com/svn/AMPS/branches/3.5-stable@129572 2c54a935-e501-0410-bc05-97a93f6bca70

+++ maven-amps-plugin/src/test/java/com/atlassian/maven/plugins/amps/TestMavenGoalsHomeZip.java
@@ -64,7 +64,6 @@ public class TestMavenGoalsHomeZip
         //when mocking the session
         //MavenSession session = mock(MavenSession.class);
 
-        PluginManager pluginManager = mock(PluginManager.class);
         List<MavenProject> reactor = Collections.<MavenProject>emptyList();
         ctx = mock(MavenContext.class);
 
@@ -90,7 +89,10 @@ public class TestMavenGoalsHomeZip
             }
             zip = null;
         }
-        FileUtils.deleteDirectory(tempDir);
+        if (tempDir != null)
+        {
+            FileUtils.deleteDirectory(tempDir);
+        }
     }
 
     @Test
